refactor: move cron-job enum
