+++ packages/apps/reputation-oracle/server/src/modules/cron-job/cron-job.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 
+import { AbuseModule } from '../abuse/abuse.module';
 import { EscrowCompletionModule } from '../escrow-completion/escrow-completion.module';
 import { IncomingWebhookModule } from '../webhook/webhook-incoming.module';
 import { OutgoingWebhookModule } from '../webhook/webhook-outgoing.module';
 
 import { CronJobService } from './cron-job.service';
 import { CronJobRepository } from './cron-job.repository';
-import { AbuseModule } from '../abuse/abuse.module';
 
 @Module({
   imports: [
