+++ packages/apps/reputation-oracle/server/src/modules/cron-job/cron-job.service.ts
@@ -1,13 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { Cron } from '@nestjs/schedule';
 
-import { CronJobType } from '../../common/enums/cron-job';
-
 import logger from '../../logger';
+
 import { AbuseService } from '../abuse/abuse.service';
 import { EscrowCompletionService } from '../escrow-completion/escrow-completion.service';
 import { IncomingWebhookService } from '../webhook/webhook-incoming.service';
 import { OutgoingWebhookService } from '../webhook/webhook-outgoing.service';
+
+import { CronJobType } from './constants';
 import { CronJobEntity } from './cron-job.entity';
 import { CronJobRepository } from './cron-job.repository';
 
@@ -27,10 +28,8 @@ export class CronJobService {
    * Starts a new cron job of the specified type if it's not already running.
    * If a cron job of the specified type doesn't exist, creates a new one and returns it.
    * If the cron job already exists, updates its start time and clears the completion time.
-   * @param cronJobType The type of cron job to start.
-   * @returns {Promise<CronJobEntity>} A Promise containing the started or updated cron job entity.
    */
-  public async startCronJob(cronJobType: CronJobType): Promise<CronJobEntity> {
+  async startCronJob(cronJobType: CronJobType): Promise<CronJobEntity> {
     const cronJob = await this.cronJobRepository.findOneByType(cronJobType);
 
     if (!cronJob) {
@@ -39,17 +38,14 @@ export class CronJobService {
       cronJobEntity.startedAt = new Date();
       return this.cronJobRepository.createUnique(cronJobEntity);
     }
+
     cronJob.startedAt = new Date();
     cronJob.completedAt = null;
+
     return this.cronJobRepository.updateOne(cronJob);
   }
 
-  /**
-   * Checks if a cron job of the specified type is currently running.
-   * @param cronJobType The type of cron job to check.
-   * @returns {Promise<boolean>} A Promise indicating whether the cron job is running.
-   */
-  public async isCronJobRunning(cronJobType: CronJobType): Promise<boolean> {
+  async isCronJobRunning(cronJobType: CronJobType): Promise<boolean> {
     const lastCronJob = await this.cronJobRepository.findOneByType(cronJobType);
 
     if (!lastCronJob || lastCronJob.completedAt) {
@@ -57,18 +53,11 @@ export class CronJobService {
     }
 
     this.logger.info('Previous cron job is not completed yet', { cronJobType });
+
     return true;
   }
 
-  /**
-   * Marks the specified cron job entity as completed.
-   * Throws an error if the cron job entity is already marked as completed.
-   * @param cronJobEntity The cron job entity to mark as completed.
-   * @returns {Promise<CronJobEntity>} A Promise containing the updated cron job entity.
-   */
-  public async completeCronJob(
-    cronJobEntity: CronJobEntity,
-  ): Promise<CronJobEntity> {
+  async completeCronJob(cronJobEntity: CronJobEntity): Promise<CronJobEntity> {
     if (cronJobEntity.completedAt) {
       throw new Error('Cron job is already completed');
     }
@@ -77,13 +66,8 @@ export class CronJobService {
     return this.cronJobRepository.updateOne(cronJobEntity);
   }
 
-  /**
-   * Processes all pending incoming webhooks, marking them as completed upon success.
-   * Handles any errors by logging them and updating the webhook status.
-   * @returns {Promise<void>} A promise that resolves when all pending incoming webhooks have been processed.
-   */
   @Cron('*/2 * * * *')
-  public async processPendingIncomingWebhooks(): Promise<void> {
+  async processPendingIncomingWebhooks(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessPendingIncomingWebhook,
     );
@@ -107,14 +91,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Processes pending escrow completion tracking to manage escrow lifecycle actions.
-   * Checks escrow status and, if appropriate, saves results and initiates payouts.
-   * Handles errors and logs detailed messages.
-   * @returns {Promise<void>} A promise that resolves when the operation is complete.
-   */
   @Cron('*/2 * * * *')
-  public async processPendingEscrowCompletion(): Promise<void> {
+  async processPendingEscrowCompletion(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessPendingEscrowCompletionTracking,
     );
@@ -138,13 +116,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Processes paid escrow completion tracking, finalizing escrow operations if completed.
-   * Notifies oracles via callbacks, logs errors, and updates tracking status.
-   * @returns {Promise<void>} A promise that resolves when the paid escrow tracking has been processed.
-   */
   @Cron('*/2 * * * *')
-  public async processPaidEscrowCompletion(): Promise<void> {
+  async processPaidEscrowCompletion(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessPaidEscrowCompletionTracking,
     );
@@ -168,13 +141,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Processes pending outgoing webhooks, sending requests to designated URLs.
-   * Updates each webhook's status upon success, retries or logs errors as necessary.
-   * @returns {Promise<void>} A promise that resolves once all pending outgoing webhooks have been processed.
-   */
   @Cron('*/2 * * * *')
-  public async processPendingOutgoingWebhooks(): Promise<void> {
+  async processPendingOutgoingWebhooks(): Promise<void> {
     if (await this.isCronJobRunning(CronJobType.ProcessPendingOutgoingWebhook))
       return;
 
@@ -193,12 +161,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Runs processing of awaiting payouts for escrow completion.
-   * @returns {Promise<void>} A promise that resolves when the processing is finished.
-   */
   @Cron('*/2 * * * *')
-  public async processAwaitingEscrowPayouts(): Promise<void> {
+  async processAwaitingEscrowPayouts(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessAwaitingEscrowPayouts,
     );
@@ -222,12 +186,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Process a pending abuse request.
-   * @returns {Promise<void>} - Returns a promise that resolves when the operation is complete.
-   */
   @Cron('*/2 * * * *')
-  public async processAbuseRequests(): Promise<void> {
+  async processAbuseRequests(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessRequestedAbuse,
     );
@@ -249,12 +209,8 @@ export class CronJobService {
     await this.completeCronJob(cronJob);
   }
 
-  /**
-   * Process a classified abuse.
-   * @returns {Promise<void>} - Returns a promise that resolves when the operation is complete.
-   */
   @Cron('*/2 * * * *')
-  public async processClassifiedAbuses(): Promise<void> {
+  async processClassifiedAbuses(): Promise<void> {
     const isCronJobRunning = await this.isCronJobRunning(
       CronJobType.ProcessClassifiedAbuse,
     );
